fix: Do not print annotation added if it already exists
Turned annotation adding into a separate (pure) checking function to see
if we have to add an annotation, and a (side-effect) function which
actually adds it.

This way we can decouple checking and send an info to io only if it is
necessary while keeping the two tasks of checking and adding separate.